Bunch of updates

+++ src/main/java/org/ldcgc/backend/controller/users/impl/VolunteerControllerImpl.java
@@ -5,6 +5,7 @@ import org.ldcgc.backend.controller.users.VolunteerController;
 import org.ldcgc.backend.payload.dto.users.VolunteerDto;
 import org.ldcgc.backend.service.users.VolunteerService;
 import org.ldcgc.backend.util.common.EOrder;
+import org.ldcgc.backend.util.common.EVStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
@@ -29,16 +30,16 @@ public class VolunteerControllerImpl implements VolunteerController {
         return volunteerService.createVolunteer(volunteerDto);
     }
 
-    public ResponseEntity<?> listVolunteers(String builderAssistantId, String filterString, Boolean isActive, Integer pageIndex, Integer size, String sortField, EOrder order) {
-        return volunteerService.listVolunteers(builderAssistantId, filterString, isActive, pageIndex, size, sortField, order);
+    public ResponseEntity<?> listVolunteers(String builderAssistantId, String filterString, EVStatus status, Integer pageIndex, Integer size, String sortField, EOrder order) {
+        return volunteerService.listVolunteers(builderAssistantId, filterString, status, pageIndex, size, sortField, order);
     }
 
     public ResponseEntity<?> updateVolunteer(String builderAssistantId, VolunteerDto volunteerDto) {
         return volunteerService.updateVolunteer(builderAssistantId, volunteerDto);
     }
 
-    public ResponseEntity<?> deleteVolunteer(String builderAssistantId) {
-        return volunteerService.deleteVolunteer(builderAssistantId);
+    public ResponseEntity<?> deleteVolunteer(String builderAssistantId, Boolean confirmDeletion) {
+        return volunteerService.deleteVolunteer(builderAssistantId, confirmDeletion);
     }
 
     public ResponseEntity<?> uploadVolunteers(Integer groupId, MultipartFile document) {
